+++ src/main/java/nextstep/security/DaoAuthenticationProvider.java
@@ -1,4 +1,4 @@
-package nextstep.app;
+package nextstep.security;
 
 import nextstep.security.authentication.Authentication;
 import nextstep.security.authentication.AuthenticationProvider;
@@ -6,11 +6,9 @@ import nextstep.security.authentication.exception.AuthenticationException;
 import nextstep.security.user.UserDetails;
 import nextstep.security.user.UserDetailsService;
 import nextstep.security.user.UsernamePasswordAuthenticationToken;
-import org.springframework.stereotype.Component;
 
 import java.util.Objects;
 
-@Component
 public class DaoAuthenticationProvider implements AuthenticationProvider {
 
     private final UserDetailsService userDetailsService;
