+++ src/main/java/nextstep/security/ProviderManager.java
@@ -1,15 +1,13 @@
-package nextstep.app;
+package nextstep.security;
 
 import nextstep.security.authentication.Authentication;
 import nextstep.security.authentication.AuthenticationManager;
 import nextstep.security.authentication.AuthenticationProvider;
 import nextstep.security.authentication.exception.AuthenticationException;
-import org.springframework.stereotype.Component;
 
 import java.util.List;
 import java.util.Objects;
 
-@Component
 public class ProviderManager implements AuthenticationManager {
 
     private final List<AuthenticationProvider> authenticationProviders;
